make utils fuctions and fix flow when use alias
