+++ .flowconfig
@@ -7,5 +7,5 @@
 [lints]
 
 [options]
-
+module.name_mapper='^@' ->'<PROJECT_ROOT>/src/'
 [strict]
