+++ config/webpack.config.dev.js
@@ -86,7 +86,8 @@ module.exports = {
     alias: {
       // Support React Native Web
       // https://www.smashingmagazine.com/2016/08/a-glimpse-into-the-future-with-react-native-for-web/
-      'react-native': 'react-native-web'
+      'react-native': 'react-native-web',
+      '@': path.resolve('src')
     },
     plugins: [
       // Prevents users from importing files from outside of src/ (or node_modules/).
